backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
youtube-transcript-api>=1.0.0
groq>=0.4.0
pydantic>=2.10.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
